requirements.txt
discord.py==1.7.3
Flask==2.0.1
google-api-python-client==2.11.0
praw==7.3.0
youtube_dl==2021.6.6
discord-components==1.0.1
PyNaCl==1.4.0
oauth==1.0.1
mod==0.3.0
aiofiles==0.7.0
DiscordUtils==1.3.4
asyncpg==0.23.0
prsaw==0.4.0
pymongo==3.11.4
dnspython==2.1.0
lavalink==3.1.4
Pillow==8.3.1